small refactoring and outsourcing of ligic from home_page.dart
